vosk/requirements.txt
flask==3.0.3
websocket-client==1.7.0
requests==2.32.3
